projeto finalizado sem a parte de cadastro de cliente

+++ api/parking-domain/src/main/java/com/itss/parking/domain/Parking.java
@@ -61,7 +61,7 @@ public class Parking {
 	private Double value;
 	
 	@Column(name = "DURATION")
-	private Long duration;
+	private Double duration;
 	
 	@Column(name = "OCCUPIED")
 	private Boolean occupied;
